+++ src/components/molecules/Table/TableBody.tsx
@@ -1,0 +1,8 @@
+export interface TableBodyProps {
+  className?: string;
+  children?: React.ReactNode;
+}
+
+export const TableBody = ({ children, className }: TableBodyProps) => (
+  <tbody className={className}>{children}</tbody>
+);
