+++ src/components/molecules/Table/TableHeader.module.scss
@@ -1,0 +1,5 @@
+.header {
+  position: sticky;
+  top: 0;
+  z-index: 5;
+}
