+++ src/components/molecules/Table/index.module.scss
@@ -1,10 +1,10 @@
-@import '../../../styles/libs';
+@use 'src/styles/libs' as libs;
 
 .table {
-  @include fade-in(0.3s, 0.3s);
+  @include libs.fade-in(0.3s, 0.3s);
   overflow: auto;
 
-  background-color: $WHITE;
+  background-color: libs.$WHITE;
 
   text-align: left;
   > table {
@@ -20,163 +20,4 @@
     width: 100%;
     height: 100%;
   }
-}
-
-.header {
-  position: sticky;
-  top: 0;
-  z-index: 5;
-}
-
-.title {
-  vertical-align: middle;
-
-  border-right: 0 solid $BLUISH-GRAY-10;
-  border-bottom: 1px solid $BLUISH-GRAY-200;
-
-  background-color: $WHITE;
-
-  cursor: move;
-
-  transition: background-color 0.3s, box-shadow 0.3s;
-  &.hovered {
-    background-color: $BLUISH-GRAY-10;
-  }
-}
-
-.title-content {
-  box-sizing: border-box;
-  padding: 6px 16px;
-
-  white-space: pre;
-
-  transition: width 0.3s;
-}
-
-.row {
-  background-color: $WHITE;
-  &:hover {
-    > .cell {
-      background-color: $BLUISH-GRAY-10;
-    }
-  }
-}
-
-.cell {
-  vertical-align: middle;
-
-  border-right: 0 solid $BLUISH-GRAY-10;
-  border-bottom: 1px solid $BLUISH-GRAY-200;
-
-  background-color: $WHITE;
-
-  transition: background-color 0.3s, box-shadow 0.3s;
-  &.hovered {
-    background-color: $BLUISH-GRAY-10;
-  }
-  &:hover {
-    z-index: 6;
-  }
-}
-
-.drop-left,
-.drop-right {
-  transition: border-width 0.3s;
-}
-
-.drop-right {
-  border-right: 20px $BLUISH-GRAY-10 solid;
-}
-
-.drop-left {
-  border-left: 20px $BLUISH-GRAY-10 solid;
-}
-
-.cell-content-container {
-  position: relative;
-
-  &.overflow {
-    overflow: hidden;
-  }
-
-  &.not-overflow {
-    overflow: visible;
-  }
-
-  box-sizing: border-box;
-  padding: 8px 16px;
-
-  text-overflow: ellipsis;
-  white-space: nowrap;
-
-  &:has(.hovered-cell-content-wrap) {
-    overflow: unset;
-  }
-}
-
-.hidden {
-  overflow: hidden;
-
-  visibility: hidden;
-}
-
-.hovered-cell-content-wrap {
-  position: absolute;
-  top: 7px;
-  left: 8px;
-  z-index: 3;
-  display: flex;
-  overflow: hidden;
-
-  width: fit-content;
-  padding: 0px 7px;
-  border: 1px solid $BLUISH-GRAY-200;
-  border-radius: 4px;
-
-  background-color: $WHITE;
-
-  white-space: pre;
-
-  transition: background-color 0.3s, border 0.3s;
-
-  @include fade-in(0.2s);
-
-  &.has-style {
-    white-space: pre-wrap;
-  }
-}
-
-.dragging > div {
-  opacity: 0.3;
-
-  transition: opacity 0.3s;
-}
-
-.restoring > div {
-  opacity: 1;
-
-  transition: opacity 0.3s;
-}
-
-.fixed {
-  position: sticky;
-  z-index: 4;
-  &.shadow {
-    box-shadow: 4px 0px 4px rgba(0, 0, 0, 0.08);
-  }
-}
-
-.hovered-cell-content {
-  overflow-y: auto;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
-
-.copy-button-wrap {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  margin-left: 5px;
 }
